+++ SakilaAPI/Repositories/Interfaces/IActorRepository.cs
@@ -5,5 +5,5 @@ namespace SakilaAPI.Repositories.Interfaces;
 public interface IActorRepository
 {
     Task<IEnumerable<Actor>> GetActorsAsync(int page, int pageSize);
-    Task<Actor?> GetActorByIdAsync(int id);
+    Task<Actor?> GetActorByIdAsync(ushort id);
 }
